backend/requirements.txt
numpy
kafka-python
torch